Added functionality for keys and locked exits, changed everything related to weight from int to long, added dashes (-) to all the strings that only have directions (in Game and Room classes), and added an example key in items.json

+++ src/textAdventure/Inventory.java
@@ -17,7 +17,7 @@ public class Inventory {
     return maxWeight;
   }
 
-  public void updateMaxWeight(int maxWeight){
+  public void updateMaxWeight(long maxWeight){
     this.maxWeight = maxWeight;
   }
